Making Featurizer class for NNScore.
